Revert "refactor: add pallet subcommand"
This reverts commit d066526fd353a26ba70f61feb6a85e14ff667249.
