+++ src/main.rs
@@ -41,8 +41,7 @@ enum Commands {
 	/// Test a smart contract.
 	#[clap(alias = "t")]
 	Test(commands::test::TestArgs),
-    /// Add a pallet into the runtime
-	#[clap(alias = "a")]
+    /// Add a pallet to the runtime
     Add(commands::add::AddArgs),
 }
 
@@ -74,9 +73,7 @@ async fn main() -> Result<()> {
 			#[cfg(feature = "contract")]
 			commands::test::TestCommands::Contract(cmd) => cmd.execute(),
 		},
-        Commands::Add(args) => match &args.command {
-			commands::add::AddCommands::Pallet(cmd) => cmd.execute(),
-		},
+        Commands::Add(args) => args.execute(),
 	}
 }
